requirements.txt
eventlet
flask
flask-socketio
flask-wtf
mysql-connector-python
python-dotenv
pyodbc
sqlalchemy
requests
PyJWT
pymongo